pyproject.toml
[project]
name = "investment-team"
version = "0.1.0"
description = "Multi-agent investment analysis system using Agno"
requires-python = ">=3.11"
dependencies = [
    "agno",
    "google-genai",
    "tavily-python",
    "yfinance",
    "pydantic>=2.0",
    "python-dotenv",
    "sqlalchemy",
    "fastapi[standard]",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
